Extract common test behavior.
